app/requirements.txt
Flask
flask-sqlalchemy
flask-migrate
Flask-User
email-validator
sqlalchemy_utils
selenium
python-dotenv
gunicorn[gevent]
psycopg2-binary